requirements.txt
django==6.0.2
psycopg2-binary==2.9.11
djangorestframework==3.16.1
drf-nested-routers==0.95.0
django-filter==25.2
djangorestframework-simplejwt==5.5.1
opencv-python==4.13.0.92
python-dotenv==1.2.1
dlib==20.0.0
face-recognition==1.3.0
django-cors-headers==4.9.0